# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StoryWeaver",
    "description": "Production-grade FastAPI backend with auth (JWT+OAuth+RBAC) and LLM chat proxy domains.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/works": {
      "get": {
        "tags": [
          "works"
        ],
        "summary": "내 작품 목록",
        "operationId": "list_works_api_v1_works_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkResponse"
                  },
                  "type": "array",
                  "title": "Response List Works Api V1 Works Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "works"
        ],
        "summary": "작품 생성",
        "operationId": "create_work_api_v1_works_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/works/{work_id}": {
      "get": {
        "tags": [
          "works"
        ],
        "summary": "작품 단건 조회",
        "operationId": "get_work_api_v1_works__work_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "works"
        ],
        "summary": "작품 수정",
        "operationId": "update_work_api_v1_works__work_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "works"
        ],
        "summary": "작품 삭제",
        "operationId": "delete_work_api_v1_works__work_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/works/{work_id}/beat-sheet": {
      "post": {
        "tags": [
          "works"
        ],
        "summary": "비트 시트 생성 (고품질 티어)",
        "operationId": "generate_beat_sheet_api_v1_works__work_id__beat_sheet_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeatSheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeatSheetResponse": {
        "properties": {
          "beats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Beats"
          }
        },
        "type": "object",
        "required": [
          "beats"
        ],
        "title": "BeatSheetResponse",
        "description": "비트 시트 생성 응답 — 회차별 비트를 한 줄씩 담는다(예: \"1화: 발단 — ...\")."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewSummary": {
        "properties": {
          "scenes": {
            "type": "integer",
            "title": "Scenes",
            "default": 0
          },
          "states": {
            "type": "integer",
            "title": "States",
            "default": 0
          },
          "conflicts": {
            "type": "integer",
            "title": "Conflicts",
            "default": 0
          }
        },
        "type": "object",
        "title": "ReviewSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "genre": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Genre"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "style": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Style"
          }
        },
        "type": "object",
        "required": [
          "title",
          "genre",
          "style"
        ],
        "title": "WorkCreate",
        "description": "작품 생성 입력 — 프론트 ``NewWorkInput``."
      },
      "WorkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "shortLabel": {
            "type": "string",
            "title": "Shortlabel"
          },
          "genre": {
            "type": "string",
            "title": "Genre"
          },
          "subGenre": {
            "type": "string",
            "title": "Subgenre"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "style": {
            "type": "string",
            "title": "Style"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "coverTheme": {
            "type": "string",
            "title": "Covertheme"
          },
          "lastEditedLabel": {
            "type": "string",
            "title": "Lasteditedlabel"
          },
          "stats": {
            "$ref": "#/components/schemas/WorkStats"
          },
          "reviewSummary": {
            "$ref": "#/components/schemas/ReviewSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "shortLabel",
          "genre",
          "subGenre",
          "keywords",
          "style",
          "status",
          "coverTheme",
          "lastEditedLabel",
          "stats",
          "reviewSummary"
        ],
        "title": "WorkResponse",
        "description": "작품 응답 — 프론트 ``Work``의 비중첩 필드 + 파생(stats·reviewSummary·lastEditedLabel).\n\n중첩 컬렉션(chapters·entities·timeline·conflicts)은 각 도메인 엔드포인트 소관이라 제외한다."
      },
      "WorkStats": {
        "properties": {
          "chapters": {
            "type": "integer",
            "title": "Chapters",
            "default": 0
          },
          "words": {
            "type": "string",
            "title": "Words",
            "default": "0"
          },
          "wordsUnit": {
            "type": "string",
            "title": "Wordsunit",
            "default": "천자"
          },
          "characters": {
            "type": "integer",
            "title": "Characters",
            "default": 0
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          }
        },
        "type": "object",
        "title": "WorkStats"
      },
      "WorkUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "shortLabel": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Shortlabel"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "subGenre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Subgenre"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "style": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "coverTheme": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Covertheme"
          }
        },
        "type": "object",
        "title": "WorkUpdate",
        "description": "작품 수정 입력 — 모든 필드 선택(PATCH)."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}